Also persist ../shed_tools directory
This is likely the problem in #118, as ../shed_tools (relative to the galaxy
directory) is the location where repository dependencies are installed by
default. This location is controlled by shed_tool_conf.xml(.sample), but since
we don't want to move around the directories too much for existing instances,
we just add ../shed_tools to the list of directories we export/import.